+++ assets/lib/jquery-weui/src/js/select.js
@@ -122,7 +122,7 @@
       });
       self.updateInputValue(values, titles);
 
-      if(config.autoClose && !config.multi) $.closePicker();
+      if(config.autoClose && !config.multi) self.close();
     })
     .on("click", ".close-select", function() {
       self.close();
@@ -146,7 +146,7 @@
 
     var config = this.config;
 
-    var dialog = this.dialog = $.openPicker(this.getHTML(), $.proxy(this.onClose, this));
+    var dialog = this.dialog = $.openPicker(this.getHTML()); // onclose 在 Select 中处理
     
     this._bind(dialog);
 
@@ -232,29 +232,29 @@
       </div>\
       </div>',
     radioTemplate:
-      '<div class="weui_cells weui_cells_radio">\
+      '<div class="weui-cells weui-cells_radio">\
         {{#items}}\
-        <label class="weui_cell weui_check_label" for="weui-select-id-{{this.title}}">\
-          <div class="weui_cell_bd weui_cell_primary">\
+        <label class="weui-cell weui-check_label" for="weui-select-id-{{this.title}}">\
+          <div class="weui-cell__bd weui-cell_primary">\
             <p>{{this.title}}</p>\
           </div>\
-          <div class="weui_cell_ft">\
-            <input type="radio" class="weui_check" name="weui-select" id="weui-select-id-{{this.title}}" value="{{this.value}}" {{#if this.checked}}checked="checked"{{/if}} data-title="{{this.title}}">\
-            <span class="weui_icon_checked"></span>\
+          <div class="weui-cell__ft">\
+            <input type="radio" class="weui-check" name="weui-select" id="weui-select-id-{{this.title}}" value="{{this.value}}" {{#if this.checked}}checked="checked"{{/if}} data-title="{{this.title}}">\
+            <span class="weui-icon-checked"></span>\
           </div>\
         </label>\
         {{/items}}\
       </div>',
     checkboxTemplate:
-      '<div class="weui_cells weui_cells_checkbox">\
+      '<div class="weui-cells weui-cells_checkbox">\
         {{#items}}\
-        <label class="weui_cell weui_check_label" for="weui-select-id-{{this.title}}">\
-          <div class="weui_cell_bd weui_cell_primary">\
+        <label class="weui-cell weui-check_label" for="weui-select-id-{{this.title}}">\
+          <div class="weui-cell__bd weui-cell_primary">\
             <p>{{this.title}}</p>\
           </div>\
-          <div class="weui_cell_ft">\
-            <input type="checkbox" class="weui_check" name="weui-select" id="weui-select-id-{{this.title}}" value="{{this.value}}" {{#if this.checked}}checked="checked"{{/if}} data-title="{{this.title}}" >\
-            <span class="weui_icon_checked"></span>\
+          <div class="weui-cell__ft">\
+            <input type="checkbox" class="weui-check" name="weui-select" id="weui-select-id-{{this.title}}" value="{{this.value}}" {{#if this.checked}}checked="checked"{{/if}} data-title="{{this.title}}" >\
+            <span class="weui-icon-checked"></span>\
           </div>\
         </label>\
         {{/items}}\